ndn: use uint32 for VAR-NUMBER
Previously, VAR-NUMBER encoding/decoding functions use uint64 but
TlvElement saves TLV-TYPE as uint32.

Given NDN Packet Format now requires minimum size encoding for
VAR-NUMBER, TLV-TYPE is limited to uint32, and TLV-LENGTH would not
exceed 4GB for practical purposes, the (incomplete) support for
VAR-NUMBER-9 (uint64) is unnecessary, and is now removed.

+++ ndn/lp.c
@@ -3,7 +3,7 @@
 #include "tlv-encoder.h"
 
 static bool
-CanIgnoreLpHeader(uint64_t tlvType)
+CanIgnoreLpHeader(uint32_t tlvType)
 {
   return 800 <= tlvType && tlvType <= 959 && (tlvType & 0x3) == 0x0;
 }
